pyproject.toml
[tool.poetry]
name = "image-pattern"
version = "0.0.16"
description = ""
authors = ["acrius <acrius@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.6"
pillow = "^7.0"
pydantic = "^1.4"
mypy = "^0.770.0"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
coverage = "^5.0"
mypy = "^0.770.0"
ipdb = "^0.13.2"
ipython = "^7.13"
django = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
